% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoBinning.R
\name{CytoBinning}
\alias{CytoBinning}
\title{perform CytoBinning on all files in a folder}
\usage{
CytoBinning(bin.number, markers, input.dir, output.dir)
}
\arguments{
\item{bin.number}{A vector of selected number of bins}

\item{markers}{A vector contains names of markers to be used in binning}

\item{input.dir}{Directory of the folder that contains pre-processed text data files}

\item{output.dir}{Directory of the folder that contains binning results, each file stores one marker pair of a given bin number}
}
\description{
This function performs CytoBinning on all data files in a folder and output a matrix where each row is the binning result for a data file
}
