% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoBinning.R
\name{preprocess}
\alias{preprocess}
\title{perform compensation and logicle transformation}
\usage{
preprocess(input.directory, output.directory, marker.index, marker.names)
}
\arguments{
\item{input.directory}{Directory of the folder contains all fcs files}

\item{output.directory}{Directory of the folder to store all transformed data in text format}

\item{marker.index}{Indices of markers that need to be logicle transformed}

\item{marker.names}{Names of markers in the same order as the indices}
}
\description{
This function performs logicle transformation on all fcs files in a folder and outputs transforemd data in text format
}
